refactor: address pre-PR review (split docker_sandbox, error-context + event-constant fixes)
