+++ src/synthorg/observability/events/docker.py
@@ -14,3 +14,5 @@ DOCKER_CONTAINER_REMOVE_FAILED: Final[str] = "docker.container.remove_failed"
 DOCKER_CLEANUP: Final[str] = "docker.cleanup"
 DOCKER_HEALTH_CHECK: Final[str] = "docker.health_check"
 DOCKER_DAEMON_UNAVAILABLE: Final[str] = "docker.daemon.unavailable"
+DOCKER_EXEC_STREAM_CLOSE_FAILED: Final[str] = "docker.exec.stream_close_failed"
+DOCKER_EXEC_INSPECT_FAILED: Final[str] = "docker.exec.inspect_failed"
